requirements.txt
pillow
pytesseract
PyMuPDF